pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "kubepath"
version = "1.0.0"
description = "Autonomous Cloud/K8s Lateral Movement Mapper — map attack paths from compromised pods to cluster admin"
readme = "README.md"
license = {text = "MIT"}
requires-python = ">=3.11"
authors = [
    {name = "KubePath Contributors"}
]
keywords = ["kubernetes", "security", "attack-graph", "lateral-movement", "rbac", "aws", "iam"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Information Technology",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Security",
]

dependencies = [
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.34.0",
    "pydantic>=2.10.0",
    "python-dotenv>=1.0.0",
    "neo4j>=5.25.0",
    "kubernetes>=31.0.0",
    "boto3>=1.36.0",
    "httpx>=0.28.0",
    "aiofiles>=24.1.0",
    "python-multipart>=0.0.20",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.25.0",
    "ruff>=0.8.0",
]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]

[tool.ruff]
target-version = "py311"
line-length = 120
